working on stacktracing the function calls - connect on port - stuck

+++ lib/data_types/datetime2.dart
@@ -51,7 +51,7 @@ class DateTime2 extends DataType {
     timestamp += (value.nanosecondsDelta ?? 0) * pow(10, scale) as int;
     timestamp = timestamp.round();
 
-    switch (scale) {
+    switch (scale.toDouble()) {
       case 0:
       case 1:
       case 2:
